Move static PDF functions to seperate file
